+++ src/Infrastructure/ZakYip.WheelDiverterSorter.Communication/Models/ChuteAssignmentNotificationEventArgs.cs
@@ -17,7 +17,7 @@ public record ChuteAssignmentNotificationEventArgs
     /// <summary>
     /// 目标格口ID（数字ID）
     /// </summary>
-    public required int ChuteId { get; init; }
+    public required long ChuteId { get; init; }
 
     /// <summary>
     /// 通知时间
